다시시작 추가

+++ Roguelike/Assets/Prefabs/Monster/Tentacle/Script/Tentacle.cs
@@ -115,7 +115,7 @@ public class Tentacle : MonsterObj
                 monsterSpr.flipX = false;
             }
 
-            characterManager.Hit(damage);
+            characterManager.Hit((int)damage);
